# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nISO API",
    "version": "1.0.0",
    "description": "Contrato das rotas de escrita que validam o corpo com Zod.\n\nGerado dos MESMOS objetos Zod que os handlers executam — não escrito à mão.\nRota que lê o corpo sem schema não aparece aqui: são as que faltam fechar\nno item 3.3 do plano, contadas pela catraca de `test/validacao-corpo.test.ts`."
  },
  "servers": [
    {
      "url": "https://niso.ness.workers.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessao": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sessão do login"
      },
      "chaveApi": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [
    {
      "sessao": []
    },
    {
      "chaveApi": []
    }
  ],
  "paths": {
...
    "/api/v1/projects/{projectId}/vendors": {
      "post": {
        "operationId": "post_api_v1_projects__projectId__vendors",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "category": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": {}
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Payload inválido — corpo recusado pelo schema, com `details` por campo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "request_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida"
          },
          "403": {
            "description": "Sessão válida, sem acesso ao recurso ou ao projeto"
          },
          "500": {
            "description": "Erro de servidor. `request_id` correlaciona com o log estruturado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "request_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vendors/{id}": {
      "put": {
        "operationId": "put_api_v1_vendors__id_",
        "tags": [
          "vendors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": {}
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Payload inválido — corpo recusado pelo schema, com `details` por campo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "request_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sem sessão válida"
          },
          "403": {
            "description": "Sessão válida, sem acesso ao recurso ou ao projeto"
          },
          "500": {
            "description": "Erro de servidor. `request_id` correlaciona com o log estruturado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "request_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}